fix bounds check in Map::getTile and return type in generate_uniform_int
Map::getTile had an inverted condition using || instead of && which would
allow out-of-bounds access. generate_uniform_int was declared and defined
as returning float instead of int, causing implicit narrowing.

+++ src/services/RandomEngine.cpp
@@ -5,7 +5,7 @@ float RandomEngine::generate_uniform_float(float min, float max) {
 	return dist(rng);
 }
 
-float RandomEngine::generate_uniform_int(int min, int max) {
+int RandomEngine::generate_uniform_int(int min, int max) {
 	std::uniform_int_distribution<int> dist(min, max);
 	return dist(rng);
 }
